+++ src/Pipfile
@@ -12,6 +12,7 @@ sklearn = "*"
 en-core-web-sm = {file = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.0.0/en_core_web_sm-2.0.0.tar.gz"}
 googlenews = "*"
 newspaper3k = "*"
+vadersentiment = "*"
 
 [requires]
 python_version = "3.8"
